% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_generate.R
\name{gs_generate}
\alias{gs_generate}
\title{gs_generate: Refine Gene Sets Based on ssGSEA Scores}
\usage{
gs_generate(
  es_out,
  es_cutoff,
  normalized_data,
  n_cutoff = 50,
  start = 0.01,
  end = 0.6
)
}
\arguments{
\item{es_out}{A data frame of enrichment scores (cells x gene sets) from \code{run_gsea}.}

\item{es_cutoff}{A data frame containing cutoffs for enrichment scores, used to define significant cells.}

\item{normalized_data}{A matrix of normalized expression data (genes x cells) for calculating marker genes.}

\item{n_cutoff}{Integer, the maximum number of top markers to select when refining each gene set.}

\item{start}{Numeric, the starting proportion for defining significant enrichment score cells.}

\item{end}{Numeric, the ending proportion for defining significant enrichment score cells.}
}
\value{
A list where the first element is the refined gene sets, and the second element is a data frame
        of differentially expressed genes for each gene set.
}
\description{
This function refines input gene sets based on ssGSEA scores and differential expression analysis.
Gene sets are iteratively adjusted by identifying marker genes associated with significant cells
in each enrichment score set.
}
\examples{
# Example data
es_out <- data.frame(GeneSet1 = runif(10), GeneSet2 = runif(10))
es_cutoff <- data.frame(GeneSet1 = 5, GeneSet2 = 4)
normalized_data <- matrix(rnorm(1000), nrow=100, ncol=10)
gs_generate(es_out, es_cutoff, normalized_data)

}
